httpx/form: Cleanup error handling
This commit cleans up error handling, particularly in the case of the
Success path. In particular, a nil Success function now results in a
panic (previously it would cause a nil dereference). Furthermore, an
'AfterSuccess' field is added to the FormContext, indicating an error is
a result of calling the success method.