umaskfree: use lock-free approach
Previously, the umaskfree package used a lock and then set the processes
umask. This had various issues, such as requiring all I/O to go via
umaskfree. This commit updates the package to instead make an explicit
call to chmod, removing the need for a lock and restrictions on the
caller.